feat: 필터링을 해주는 로직 Hook으로 분리
